requirements.txt
pywebview==5.*
pyserial==3.*
sprotocol
